Treat fields as subset of blocks
