+++ simple-analytics.php
@@ -47,13 +47,12 @@ require __DIR__ . '/src/Settings/Blocks/CalloutBlock.php';
 require __DIR__ . '/src/Settings/Concerns/HasDocs.php';
 require __DIR__ . '/src/Settings/Concerns/HasPlaceholder.php';
 require __DIR__ . '/src/Settings/Concerns/ManagesBlocks.php';
-require __DIR__ . '/src/Settings/Concerns/ManagesFields.php';
 require __DIR__ . '/src/Settings/Page.php';
-require __DIR__ . '/src/Settings/Fields/Field.php';
-require __DIR__ . '/src/Settings/Fields/Input.php';
-require __DIR__ . '/src/Settings/Fields/Checkboxes.php';
-require __DIR__ . '/src/Settings/Fields/Checkbox.php';
-require __DIR__ . '/src/Settings/Fields/IpList.php';
+require __DIR__ . '/src/Settings/Blocks/Fields/Field.php';
+require __DIR__ . '/src/Settings/Blocks/Fields/Input.php';
+require __DIR__ . '/src/Settings/Blocks/Fields/Checkboxes.php';
+require __DIR__ . '/src/Settings/Blocks/Fields/Checkbox.php';
+require __DIR__ . '/src/Settings/Blocks/Fields/IpList.php';
 require __DIR__ . '/src/Settings/Tab.php';
 require __DIR__ . '/src/UI/LabelComponent.php';
 require __DIR__ . '/src/UI/TabListComponent.php';
