+++ src/Settings/Blocks/Fields/Field.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SimpleAnalytics\Settings\Fields;
+namespace SimpleAnalytics\Settings\Blocks\Fields;
 
 use SimpleAnalytics\Settings\Blocks\Block;
